refactor(tests): ♻️ split HACS-only tests into dedicated *_custom.py files
- move v1→v4 async_migrate_entry scenarios + legacy data→options test
  out of test_init.py and into test_init_custom.py (8 tests, 358 lines)
- move vistapool import flow + v1 unmigrated duplicate abort tests out
  of test_config_flow.py and into test_config_flow_custom.py (8 tests)
- the framework-style halves of the original files now read 1:1 with
  what core would publish — no marker noise, no conditional imports
- the sync script can simply exclude the *_custom.py files via
  EXCLUDE_TEST_FILES instead of stripping interleaved blocks
- 258 tests still pass, 100 % coverage maintained

Why: marker-based stripping forced custom-only imports (DEFAULT_PORT,
async_migrate_entry, etc.) into the main import block of otherwise
framework-shaped files, then required the sync script to surgically
delete fragments of an import statement. Dedicated files make the
boundary visible from the file tree and reduce sync to a flat exclude.
